chore: update vitest to v4 and cleanups
