+++ package.json
@@ -663,7 +663,7 @@
     "@types/node": "^24.3.0",
     "@types/ws": "^8.18.1",
     "@typescript/native-preview": "7.0.0-dev.20260210.1",
-    "@vitest/coverage-v8": "^3.2.4",
+    "@vitest/coverage-v8": "^4.1.7",
     "arg": "^5.0.2",
     "bun-types": "^1.2.20",
     "editorconfig-checker": "6.1.1",
@@ -680,7 +680,7 @@
     "typescript": "^5.9.2",
     "undici": "^6.21.3",
     "vite-plugin-fastly-js-compute": "^0.4.2",
-    "vitest": "^3.2.4",
+    "vitest": "^4.1.7",
     "wrangler": "4.12.0",
     "ws": "^8.18.0",
     "zod": "^3.23.8"
