+++ src/middleware/cache/index.test.ts
@@ -1,8 +1,41 @@
 import type { ExecutionContext } from '../../context'
+import { vi } from 'vitest'
 import { Hono } from '../../hono'
 import { cache } from '.'
 
 // Mock
+type StoreMap = Map<string | Request, Response>
+
+class MockCache {
+  name: string
+  store: StoreMap
+
+  constructor(name: string, store: StoreMap) {
+    this.name = name
+    this.store = store
+  }
+
+  async match(key: Request | string): Promise<Response | null> {
+    return this.store.get(key) || null
+  }
+
+  async keys() {
+    return this.store.keys()
+  }
+
+  async put(key: Request | string, response: Response): Promise<void> {
+    this.store.set(key, response)
+  }
+}
+
+const globalStore: StoreMap = new Map()
+
+vi.stubGlobal('caches', {
+  open: (name: string) => {
+    return new MockCache(name, globalStore)
+  },
+})
+
 class Context implements ExecutionContext {
   passThroughOnException(): void {
     throw new Error('Method not implemented.')
